Fix CWE-190 map header length overflow
Promote map entry count multiplication to long when validating the minimum encoded payload length and when skipping map contents. This keeps oversized malformed map headers on the insufficient-buffer path instead of relying on Int32 arithmetic behavior. Add regression coverage for oversized map headers in ReadMapHeader and Skip.

+++ src/MessagePack/MessagePackReader.cs
@@ -371,7 +371,7 @@ namespace MessagePack
             // Protect against corrupted or mischievous data that may lead to allocating way too much memory.
             // We allow for each primitive to be the minimal 1 byte in size, and we have a key=value map, so that's 2 bytes.
             // Formatters that know each element is larger can optionally add a stronger check.
-            ThrowInsufficientBufferUnless(this.reader.Remaining >= count * 2);
+            ThrowInsufficientBufferUnless(this.reader.Remaining >= (long)count * 2);
 
             return count;
         }
